+++ docs/changes.rst
@@ -10,6 +10,10 @@ To be released.
   <libearth.subscribe.SubscriptionSet.contains>` method which provides
   ``recursively=True`` option.  It's useful for determining that
   a subcategory or subscription is in the whole tree.
+- :attr:`Attribute.default <libearth.schema.Attribute.default>` option
+  becomes to accept only callable objects.  Below 0.2.0,
+  :attr:`~libearth.schema.Attribute.default` is not a function but a value
+  which is simply used as it is.
 
 
 Version 0.1.0